requirements.txt
python-binance==1.0.12
backtrader==1.9.76.123
matplotlib--3.2.2